Moved cdc to Sofast
