+++ opencsp/app/sofast/lib/calculation_data_classes.py
